requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
httpx==0.28.1
pyrogram==2.0.106
openai==1.102.0
python-dotenv==1.1.1
psycopg2-binary==2.9.10
